Move some types to fontdrasil. No edits, just move.
