+++ fontbe/src/fvar.rs
@@ -93,7 +93,8 @@ impl Work<Context, Error> for FvarWork {
 
 #[cfg(test)]
 mod tests {
-    use fontir::ir::{Axis, StaticMetadata};
+    use fontdrasil::axis::Axis;
+    use fontir::ir::StaticMetadata;
 
     use super::generate_fvar;
