+++ fontdrasil/Cargo.toml
@@ -11,11 +11,17 @@ categories = ["text-processing", "parsing", "graphics"]
 [features]
 
 [dependencies]
-smol_str.workspace = true
+font-types.workspace = true
 
+smol_str.workspace = true
+ordered-float.workspace = true
 serde.workspace = true
+
+log.workspace = true
 
 [dev-dependencies]
 diff.workspace = true
 ansi_term.workspace = true
+kurbo.workspace = true
 tempfile.workspace = true
+pretty_assertions.workspace = true
