+++ fontdrasil/src/lib.rs
@@ -1,5 +1,11 @@
 //! Helper library for code that is of value to FE and BE of font compilation.
 
+pub mod axis;
+pub mod coords;
+pub mod error;
 pub mod orchestration;
 pub mod paths;
+pub mod piecewise_linear_map;
+pub(crate) mod serde;
 pub mod types;
+pub mod variations;
