+++ glyphs2fontir/src/toir.rs
@@ -1,9 +1,12 @@
 use std::{collections::HashMap, str::FromStr};
 
 use font_types::Tag;
-use fontdrasil::types::GlyphName;
-use fontir::{
+use fontdrasil::{
+    axis::Axis,
     coords::{CoordConverter, DesignCoord, DesignLocation, NormalizedLocation, UserCoord},
+    types::GlyphName,
+};
+use fontir::{
     error::{Error, WorkError},
     ir::{self, GlyphPathBuilder},
 };
@@ -116,7 +119,7 @@ pub(crate) fn to_ir_features(features: &[FeatureSnippet]) -> Result<ir::Features
     Ok(ir::Features::Memory(fea_snippets.join("\n\n")))
 }
 
-fn design_location(axes: &[ir::Axis], axes_values: &[OrderedFloat<f64>]) -> DesignLocation {
+fn design_location(axes: &[Axis], axes_values: &[OrderedFloat<f64>]) -> DesignLocation {
     axes.iter()
         .zip(axes_values.iter())
         .map(|(axis, pos)| (axis.name.clone(), DesignCoord::new(pos.into_inner() as f32)))
@@ -148,7 +151,7 @@ fn to_ir_axis(
     axis_values: &[OrderedFloat<f64>],
     default_idx: usize,
     axis: &glyphs_reader::Axis,
-) -> Result<ir::Axis, Error> {
+) -> Result<Axis, Error> {
     let min = axis_values
         .iter()
         .map(|v| OrderedFloat::<f32>(v.into_inner() as f32))
@@ -187,7 +190,7 @@ fn to_ir_axis(
         CoordConverter::unmapped(min, default, max)
     };
 
-    Ok(ir::Axis {
+    Ok(Axis {
         name: axis.name.clone(),
         tag: Tag::from_str(&axis.tag).map_err(Error::InvalidTag)?,
         hidden: axis.hidden.unwrap_or(false),
@@ -198,7 +201,7 @@ fn to_ir_axis(
     })
 }
 
-fn ir_axes(font: &Font) -> Result<Vec<ir::Axis>, Error> {
+fn ir_axes(font: &Font) -> Result<Vec<Axis>, Error> {
     // Every master should have a value for every axis
     for master in font.masters.iter() {
         if font.axes.len() != master.axes_values.len() {
@@ -226,7 +229,7 @@ pub struct FontInfo {
     pub master_indices: HashMap<String, usize>,
     /// Axes values => location for every instance and master
     pub locations: HashMap<Vec<OrderedFloat<f64>>, NormalizedLocation>,
-    pub axes: Vec<ir::Axis>,
+    pub axes: Vec<Axis>,
     /// Index by tag
     pub axis_indices: HashMap<Tag, usize>,
 }
